# solver warmstart