@thowell 2025-06-01
@erikfrey the pr is updated with a simplified warmstart option for the solver