@thowell 2025-07-24
@erikfrey the changes in this pr are now quite minimal, essentially just an option to initialize the solve `qacc` with `qacc_smooth` instead of `qacc_warmstart`. my preference would be to merge this pr as-is and then investigate solver iterations for difference scenarios in the future.